+++ Documentation/Changelog.md
@@ -26,6 +26,11 @@
   schemarouter clear <service>` where `<service>` is a service that uses the
   schemarouter.
 
+* Schemarouter now allows stale cache entries to be used while the database map
+  cache is being updated. By default the entries are usable for 150 seconds
+  after they have gone stale. This limit can be configured with the
+  `max_staleness` parameter.
+
 ## MariaDB MaxScale 23.02
 
 * A transition from the traditional _master/slave_ terminology to the